refactor: move logic to ProfileSocialCounters component
