+++ src/app/(marketing)/profile/_components/ProfileSocialCounters.tsx
@@ -1,0 +1,104 @@
+import FarcasterImage from "@/../static/socialIcons/farcaster.svg"
+import { AvatarGroup } from "@/components/ui/AvatarGroup"
+import { Separator } from "@/components/ui/Separator"
+import { cn } from "@/lib/utils"
+import { FarcasterProfile } from "@guildxyz/types"
+import useUser from "components/[guild]/hooks/useUser"
+import useSWRImmutable from "swr/immutable"
+import { useProfile } from "../_hooks/useProfile"
+import { useReferredUsers } from "../_hooks/useReferredUsers"
+
+export const ProfileSocialCounters = ({ className }: any) => {
+  const { data: referredUsers } = useReferredUsers()
+  const profileFarcaster = useProfileFarcaster()
+
+  const fc = profileFarcaster.followers.data?.users.at(0)
+  const relevantFc =
+    profileFarcaster.relevantFollowers.data?.top_relevant_followers_hydrated?.map(
+      ({ user }) => user
+    )
+
+  return (
+    <div
+      className={cn(
+        "flex flex-wrap items-center justify-center gap-6 gap-y-4 sm:flex-nowrap",
+        className
+      )}
+    >
+      <div className="flex flex-col items-center leading-tight">
+        <div className="font-bold md:text-lg">{referredUsers.length}</div>
+        <div className="text-muted-foreground">Guildmates</div>
+      </div>
+      {fc && (
+        <>
+          <Separator orientation="vertical" className="h-10 md:h-12" />
+          <div className="flex flex-col items-center leading-tight">
+            <div className="font-bold md:text-lg">{fc.following_count}</div>
+            <div className="flex items-center gap-1 text-muted-foreground">
+              <FarcasterImage />
+              Following
+            </div>
+          </div>
+
+          <Separator
+            orientation="horizontal"
+            className="h-px w-full sm:h-10 sm:w-px md:h-12"
+          />
+          {relevantFc ? (
+            <RelevantFollowers {...{ relevantFc, fc }} />
+          ) : (
+            <div className="flex flex-col items-center leading-tight">
+              <div className="font-bold md:text-lg">{fc.follower_count}</div>
+              <div className="flex items-center gap-1 text-muted-foreground">
+                <FarcasterImage />
+                Followers
+              </div>
+            </div>
+          )}
+        </>
+      )}
+    </div>
+  )
+}
+
+const useProfileFarcaster = () => {
+  const { data: profile } = useProfile()
+  const user = useUser(profile?.userId)
+  const userFcProfile = user.farcasterProfiles?.at(0)
+  const fcProfile = useSWRImmutable<FarcasterProfile[]>(
+    profile?.userId ? `/v2/users/${profile.userId}/farcaster-profiles` : null
+  ).data?.at(0)
+  const followers = useSWRImmutable(
+    fcProfile
+      ? `https://api.neynar.com/v2/farcaster/user/bulk?api_key=NEYNAR_API_DOCS&fids=${fcProfile.fid}`
+      : null
+  )
+  const relevantFollowers = useSWRImmutable(
+    fcProfile && userFcProfile
+      ? `https://api.neynar.com/v2/farcaster/followers/relevant?api_key=NEYNAR_API_DOCS&target_fid=${fcProfile.fid}&viewer_fid=${userFcProfile.fid}`
+      : null
+  )
+  return { followers, relevantFollowers }
+}
+
+const RelevantFollowers = ({ relevantFc, fc }: { relevantFc: any; fc: any }) => {
+  const [first, second] = relevantFc
+  return (
+    <div className="flex items-center gap-2">
+      <AvatarGroup
+        imageUrls={relevantFc.map(({ pfp_url }) => pfp_url)}
+        count={fc.follower_count}
+      />
+      <div className="max-w-64 text-muted-foreground leading-tight">
+        Followed by <span className="font-bold">{first.display_name}</span>
+        {second && (
+          <>
+            <span>", "</span>
+            <span className="font-bold">{second.display_name}</span>
+          </>
+        )}{" "}
+        and {fc.follower_count - Math.min(2, relevantFc.length)} others on Farcaster
+      </div>
+    </div>
+  )
+}
